requirements.txt
duckdb
pandas
pyarrow
numpy
openpyxl
plotly
streamlit
pyyaml
rapidfuzz
